pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pandas-stubs>=2.3.0.250703",
    "pretty-errors>=1.2.25",
    "python-dotenv>=1.1.1",
    "sqlglot[rs]>=27.0.0",
]

[tool.uv]
managed = true
